pyproject.toml
[project]
name = "irimo-telbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "pytelegrambotapi>=4.29.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
